requirements.txt
pandas==2.1.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
matplotlib==3.8.2
seaborn==0.13.0
pyyaml==6.0.1
pytest==7.4.3
python-dotenv==1.0.0
jupyter==1.0.0
ipykernel==6.27.1
